Rename env variable, refactor tests
Drops the ancient rc-file convention in favor of YAMLLINT_CONFIG_FILE to
be more consistent with the rest of the documentation.

Uses the `tempfile` module's `NamedTemporaryFile` context manager to
handle the creation and cleanup of temp files (`NamedTemporaryFile` has
a `delete` option, true by default, which removes the file once it's
closed).