Document that GitHub Actions already installs `yamllint`
I was looking up instructions for how to install `yamllint` so that I could use it in our GitHub actions workflow when I happened to see mention that it's already installed in the base image.

So adding a quick note here.

I phrased the note generically because additional CI systems may (should!! 😄 ) start doing this, so we can easily add them to the list.