Add (optional) multi-processing functionality
Add a command-line option to allow an end-user to specify a number of
separate worker processes to use, in order to potentially speed up
linting of large repositories.

Default the option to 1 to (largely?) preserve the current behavior.
Possibly worth noting that, technically, we are still multiprocessing
with a value of '1', as we still spin up a separate worker process from
the main process and then just wait on that process to finish the work.