Exit immediately on error unless specified
This commit restores the default functionality of exiting immediately
with error code -1 if an error occurs while parsing files. Also add a
--continue-on-error flag in case users want to continue parsing any
remaining files before exiting.

This could be convenient for large batch jobs, broken symlinks, etc., by
allowing the user to still get the linting output even if one of the
files is missing, not readable, or similar.