config: Report if 'rules' is not a dict
In the current version, the error message states `TypeError: 'NoneType' object
is not iterable` when the 'rules' in the config file is not a dict (gh-274.) It
would be helpful to improve the error message.

Additionally, it is for consistency, as other elements (`ignore`, `locale`,
etc.) have similar error checks.